Add strictness tests for Map construction
This aims to reduce the chance of introducing strictness bugs.
Since we use the same Map type for lazy and strict maps, it is not
possible to ensure appropriate strictness at the type level. So we turn
to property tests.

Arbitrary Set and Map generation is moved from set-properties.hs and
map-properties.hs to ArbitrarySetMap.hs to be shared with the new
strictness tests.