Use randomly structured Set and Map in benchmarks (#1075)
Trees constructed from [1..n] have a specific structure of perfect
binary trees linked together. We shuffle the list so that the tree
is generated from repeated insertions instead, which forms a random
structure that should be more representative of average use cases.

Most benchmarks show an increase in measured times, the amount varying
from 10% to 70%. This is expected because the trees are more unbalanced
now. A handful of benchmarks show larger increases. These were previously
measuring best-case scenarios, which no longer occur with random trees.