In docs, use "look up" instead of "lookup" when used as a verb.
"To look up (a thing)" is a phrasal verb, whereas the single word
"lookup" (written without spaces) is generally used as a noun (for
example: "in order to avoid an expensive lookup") and occasionally
as an adjective (for example: "a lookup table").

Some parts of the documentation already use "look up" in a verbal
context.

For example, `Data.Map.Internal.hs:1295` models correct usage:

```hs
-- Look up a key and return a result indicating whether it was found
-- and what path was taken.
lookupTrace :: Ord k => k -> Map k a -> TraceResult a
```

This commit adjusts the rest of the documentation (and comments) to
match this usage (when "look up" is used as a verb). When "lookup"
is used as a noun or an adjective, we leave it alone.

This usage of phrasal verbs is similar to:

- "pick up"  (verb) vs "pickup"  (noun)
- "sit up"   (verb) vs "situp"   (noun)
- "log in"   (verb) vs "login"   (noun)
- "start up" (verb) vs "startup" (noun)
- "set up"   (verb) vs "setup"   (noun)