Add link{L,R} for Set and Map
linkL checks whether the left tree is heavier than the right tree, but
unlike link it skips checking the other way around. linkR checks the
opposite.

These are used where we know for certain that we only need to rebalance
in one direction. link also delegates to these since the direction never
changes during linking. L and R versions for merge/link2 are added for
the same reason.