Make Int{Map,Set} folds friendlier to optimizations
Move the Nil branch to the top-level and error on Nil in go.
The Note [IntMap folds] added in Data.IntMap.Internal explains the
details.

For the "maximum" benchmarks added here, the time improves by ~40% for
IntMap and ~30% for sparse IntSets. "traverseSum" and "cpsSum" benchmarks
also improve by 30-60% for IntMaps and sparse IntSets. Dense IntSets are
barely affected.