Drop "ifdef GHC"s on INLINABLE pragmas
These were added long ago in 06e91acc as an improvement on GHC over the
already existing INLINE pragmas. We can get rid of these unnecessary
(and confusing to contributors) ifdefs today since we do not have other
compilers where we want the non-INLINABLE behavior to apply.

While INLINABLE is not a standard pragma listed in the Haskell 2010
report, the report also specifies that unrecognized pragmas should be
ignored. So we don't expect this to cause issues for non-GHC compilers.
MicroHs quietly ignores them as expected.